Move DescribeAddressVisitor to rpc/util
Modify for BLS
