+++ src/rpc/misc.cpp
@@ -31,118 +31,6 @@
 #include <malloc.h>
 #endif
 
-#include <wallet/wallet.h>
-
-class DescribeAddressVisitor : public std::variant<UniValue> {
-public:
-    explicit DescribeAddressVisitor() {}
-
-    UniValue operator()(const CNoDestination &dest) const {
-        return UniValue(UniValue::VOBJ);
-    }
-
-    UniValue operator()(const CKeyID &keyID) const {
-        UniValue obj(UniValue::VOBJ);
-        obj.pushKV("isscript", false);
-        return obj;
-    }
-
-    UniValue operator()(const CScriptID &scriptID) const {
-        UniValue obj(UniValue::VOBJ);
-        obj.pushKV("isscript", true);
-        return obj;
-    }
-};
-
-#ifdef ENABLE_WALLET
-class DescribeWalletAddressVisitor : public std::variant<UniValue> {
-public:
-    CWallet *const pwallet;
-
-    void ProcessSubScript(const CScript &subscript, UniValue &obj,
-                          bool include_addresses = false) const {
-        // Always present: script type and redeemscript
-        txnouttype which_type;
-        std::vector<std::vector<uint8_t>> solutions_data;
-        Solver(subscript, which_type, solutions_data);
-        obj.pushKV("script", GetTxnOutputType(which_type));
-        obj.pushKV("hex", HexStr(subscript.begin(), subscript.end()));
-
-        CTxDestination embedded;
-        UniValue a(UniValue::VARR);
-        if (ExtractDestination(subscript, embedded)) {
-            // Only when the script corresponds to an address.
-            UniValue subobj(UniValue::VOBJ);
-            UniValue detail = std::visit(DescribeAddressVisitor(), embedded);
-            subobj.pushKVs(detail);
-            UniValue wallet_detail = std::visit(*this, embedded);
-            subobj.pushKVs(wallet_detail);
-            subobj.pushKV("address", EncodeDestination(embedded));
-            subobj.pushKV("scriptPubKey",
-                          HexStr(subscript.begin(), subscript.end()));
-            // Always report the pubkey at the top level, so that
-            // `getnewaddress()['pubkey']` always works.
-            if (subobj.exists("pubkey")) {
-                obj.pushKV("pubkey", subobj["pubkey"]);
-            }
-            obj.pushKV("embedded", std::move(subobj));
-            if (include_addresses) {
-                a.push_back(EncodeDestination(embedded));
-            }
-        } else if (which_type == TX_MULTISIG) {
-            // Also report some information on multisig scripts (which do not
-            // have a corresponding address).
-            // TODO: abstract out the common functionality between this logic
-            // and ExtractDestinations.
-            obj.pushKV("sigsrequired", solutions_data[0][0]);
-            UniValue pubkeys(UniValue::VARR);
-            for (size_t i = 1; i < solutions_data.size() - 1; ++i) {
-                CPubKey key(solutions_data[i].begin(), solutions_data[i].end());
-                if (include_addresses) {
-                    a.push_back(EncodeDestination(key.GetKeyID()));
-                }
-                pubkeys.push_back(HexStr(key.begin(), key.end()));
-            }
-            obj.pushKV("pubkeys", std::move(pubkeys));
-        }
-
-        // The "addresses" field is confusing because it refers to public keys
-        // using their P2PKH address. For that reason, only add the 'addresses'
-        // field when needed for backward compatibility. New applications can
-        // use the 'pubkeys' field for inspecting multisig participants.
-        if (include_addresses) {
-            obj.pushKV("addresses", std::move(a));
-        }
-    }
-
-    explicit DescribeWalletAddressVisitor(CWallet *_pwallet)
-        : pwallet(_pwallet) {}
-
-    UniValue operator()(const CNoDestination &dest) const {
-        return UniValue(UniValue::VOBJ);
-    }
-
-    UniValue operator()(const CKeyID &keyID) const {
-        UniValue obj(UniValue::VOBJ);
-        CPubKey vchPubKey;
-        if (pwallet && pwallet->GetPubKey(keyID, vchPubKey)) {
-            obj.pushKV("pubkey", HexStr(vchPubKey));
-            obj.pushKV("iscompressed", vchPubKey.IsCompressed());
-        }
-        return obj;
-    }
-
-    UniValue operator()(const CScriptID &scriptID) const {
-        UniValue obj(UniValue::VOBJ);
-        CScript subscript;
-        if (pwallet && pwallet->GetCScript(scriptID, subscript)) {
-            ProcessSubScript(subscript, obj, true);
-        }
-        return obj;
-    }
-};
-#endif
-
 #ifndef ENABLE_WALLET
 /**
  * @note Do not add or change anything in the information returned by this
